add celery task for getting intersection features
